(MoceanAPI/MIP#4) Change Mccc to Mc
