+++ src/test/java/com/mocean/modules/voice/mc/DialTest.java
@@ -1,4 +1,4 @@
-package com.mocean.modules.voice.mccc;
+package com.mocean.modules.voice.mc;
 
 import com.mocean.exception.RequiredFieldException;
 import org.junit.jupiter.api.Test;
@@ -9,32 +9,35 @@ import java.util.HashMap;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class SleepTest {
+public class DialTest {
     @Test
     public void testRequestParams() throws RequiredFieldException {
         HashMap<String, Object> params = new HashMap<String, Object>() {{
-            put("duration", 10000);
-            put("action", "sleep");
+            put("to", "testing to");
+            put("action", "dial");
+            put("from", "callerid");
+            put("dial-sequentially", true);
         }};
-        Sleep sleep = new Sleep(params);
+        Dial dial = new Dial(params);
 
-        assertEquals(params, sleep.getRequestData());
+        assertEquals(params, dial.getRequestData());
 
-        sleep = new Sleep();
-        sleep.setDuration(10000);
+        dial = new Dial();
+        dial.setTo("testing to");
+        dial.setFrom("callerid");
+        dial.setDialSequentially(true);
 
-        assertEquals(params, sleep.getRequestData());
+        assertEquals(params, dial.getRequestData());
     }
 
     @Test
     public void testIfActionAutoDefined() throws RequiredFieldException {
         HashMap<String, Object> params = new HashMap<String, Object>() {{
-            put("duration", 10000);
-            put("barge-in", true);
+            put("to", "testing to");
         }};
-        Sleep sleep = new Sleep(params);
+        Dial dial = new Dial(params);
 
-        assertEquals("sleep", sleep.getRequestData().get("action"));
+        assertEquals("dial", dial.getRequestData().get("action"));
     }
 
     @Test
@@ -42,8 +45,8 @@ public class SleepTest {
         assertThrows(RequiredFieldException.class, new Executable() {
             @Override
             public void execute() throws Throwable {
-                Sleep sleep = new Sleep();
-                sleep.getRequestData();
+                Dial dial = new Dial();
+                dial.getRequestData();
             }
         });
     }
